Move disease components to separate directory
