+++ src/App.js
@@ -5,7 +5,7 @@ import { OtUiThemeProvider } from 'ot-ui';
 
 import HomePage from './public/home/Page';
 import TargetPage from './public/target/Page';
-import DiseasePage from './pages/DiseasePage';
+import DiseasePage from './public/disease/Page';
 
 class App extends Component {
   render() {
