+++ src/public/common/ScoreCell.js
