+++ src/public/disease/Associations.js
