+++ src/public/disease/DiseaseSummary.js
